pyproject.toml
[project]
name = "mcp-client-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.50.0",
    "beautifulsoup4>=4.13.4",
    "boto3>=1.42.2",
    "fastapi>=0.115.12",
    "flask>=3.1.0",
    "jinja2>=3.1.5",
    "mcp>=1.6.0",
    "openai>=2.8.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
    "gunicorn>=21,<23",
    "pytest~=8.4.2",
    "pytest-cov==5.0.0",
    "coverage==7.6.0",
    "pydantic>=2.11.2",
    "pydantic-settings>=2.8.1",
    "pydantic-core>=2.33.1",
    "uvicorn>=0.34.0"
]
